+++ crates/contract/src/foreign_chains_metadata.rs
